processor/requirements.txt
asyncio-nats-client
git+https://github.com/roguh/newspaper.git#egg=newspaper3k # newspaper3k
aiohttp[speedups]
brotlipy
asyncpg
aiohttp
python-dotenv
uvloop
